Refactor Shift class Id property and add constructor
Updated the `Id` property in the `Shift` class to allow external assignment by removing the default GUID assignment. Introduced a constructor to initialize the `Id` with a new GUID upon instantiation. Included the `System.ComponentModel.DataAnnotations.Schema` namespace to enable the use of the `[ForeignKey]` attribute for the `UserId` property.